recovery: Prevent mounting system image if MDTP is activated
When MDTP (Mobile Device Theft Protection) is activated, mounting
the system image is a security risk that must be blocked, as it allows
a user to run adb shell.
Also, mounting the system image as RW will modify the system image
and will be detected by MDTP as an unauthorized attack.

Change-Id: I57097441455d9eb8eb8f966e1c53b74b6a0854a8